+++ tests/live-example.rs
@@ -1,13 +1,13 @@
 //! The example from README.md. Having it here ensures it compiles.
-use std::{collections::HashMap, error::Error};
+use std::error::Error;
 
 use databento::{
     dbn::{
         datasets,
         enums::{SType, Schema},
-        record::{SymbolMappingMsg, TradeMsg},
+        record::TradeMsg,
     },
-    live::Subscription,
+    live::{Subscription, SymbolMap},
     LiveClient,
 };
 
@@ -30,23 +30,16 @@ async fn main() -> Result<(), Box<dyn Error>> {
         .unwrap();
     client.start().await?;
 
-    let mut symbol_mappings = HashMap::new();
+    let mut symbol_map = SymbolMap::new();
     // Get the next trade
     loop {
         let rec = client.next_record().await?.unwrap();
-        if rec.has::<TradeMsg>() {
-            let trade = rec.get::<TradeMsg>().unwrap();
-            let symbol = symbol_mappings.get(&trade.hd.instrument_id).unwrap();
+        if let Some(trade) = rec.get::<TradeMsg>() {
+            let symbol = &symbol_map[trade.hd.instrument_id];
             println!("Received trade for {symbol}: {trade:?}",);
             break;
-        } else if rec.has::<SymbolMappingMsg>() {
-            let sym_map = rec.get::<SymbolMappingMsg>().unwrap();
-            symbol_mappings.insert(
-                sym_map.hd.instrument_id,
-                sym_map.stype_out_symbol()?.to_owned(),
-            );
         }
+        symbol_map.on_record(rec)?;
     }
-
     Ok(())
 }